move plugin to pkg/plugin
